kitchensink dependency, and bring in testutils for logging
We changed the name of the clj-utils library to `kitchensink`.
This commit changes our dependency accordingly.

We also decided to pull logging functionality out of kitchensink
and move it directly into trapperkeeper.  This commit brings
in the logging *testutils*, which are needed for our tests;
however, it doesn't bring in any of the other logging functionality
just yet.

Conflicts:

	test/puppetlabs/trapperkeeper/core_test.clj