(PE-2178) Fix and tests for Jetty SSL support
There was an extra and invalid line of code in the jetty config that
would be executed when using a pem-based SSL configuration. This line
threw an exception under Jetty 7 (as it doesn't exists in 7) and would
prevent server initialization. This commit removes that line of code
and verifies that we can initialize the webserver when using both a
pem and jks style configuration.