Remove puppet and puppetdb-specific strings
There were some strings (used in log messages, docs, ssl certificate
aliases, etc.) that referred to PuppetDB.  I believe that these
things don't actually have any tight coupling with Puppet or PuppetDB,
so I've changed them to make them more general.

It's possible that we'll need to make the SSL certificate alias
configurable, if PuppetDB is expecting for it to exactly match
the value that it was previously using ("PuppetDB CA"), but
I've not done that yet because I'm not sure whether it's actually
necessary.