Fix bug that would allow shutdown to be run twice
This commit makes a slight change to the behavior of the shutdown
hook code so that it checks to see if the `shutdown-reason` promise
has been realized before it calls the shutdown logic.

Without this change, in the case where a shutdown was triggered
by a user request or an error, the `run` function would call
the shutdown logic and then the main thread would exit, which
would trigger the JVM shutdown hook, which would then execute
the shutdown code a second time.