Supply context to shutdown-on-error error fn
Prior to this commit, if you registered an on-error handler with
the `shutdown-on-error` function from the shutdown service,
the on-error fn had to be a no-arg fn.  This gave you no way
to access your service's context if you needed to.

This commit changes the behavior so that the on-error fn must
be a one-arg fn that accepts the service context map as its
argument.