(PE-2691) allow servlet init params to be plumbed into add-servlet-handler
Prior to this commit, there was no way for a trapperkeeper service using
the add-servlet-handler function from the jetty web service to be able
to plumb through servlet init params such that the running servlet could
read its parameters from the ServletConfig it receives from Jetty during
initialization.  This commit adds an optional map parameter to the
add-servlet-handler which, if specified, is set onto the Jetty servlet's
ServletHolder.  This allows Jetty to make the resulting init parameters
available to the servlet through the ServletConfig.