Add lifecycle fns to schema for service fnks
Prior to this commit, the output schema for the service fnks was
only including the functions from their protocols; it did not
include the lifecycle functions.  While incorrect, this was not
causing test failures because the default behavior for the
prismatic graph library is to *not* enforce validation of the
fnk schemas; referencing the lifecycle functions worked fine
with schema validation disabled since they were present in
the fnk map, just not in the output schema.

Failures did show up in other code bases where the validation of
schemas had been globally enabled, which highlighted the fact that
we should be using the schema validation test fixture in our tests.

This commit adds that test fixture, which did indeed cause the
tests to start failing; it also fixes the broken schema to get
the tests passing.

We also have been forced to temporarily use a forked version of
prismatic plumbing, due to a bug that makes their current
release incompatible with PuppetDB's aot.