Change deploy method to use `lein-deploy`
Using the ":clojars" method of deployment with the lein release
plugin does not result in the inclusion of test jars.  This
commit changes the deploy method to ":lein-deploy", which we have
had success with in the past.

This approach requires you to set the environment variables
"CLOJARS_USERNAME" and "CLOJARS_PASSWORD" before running
`lein release`.