(PE-2843) Make it possible to get a reference to a service from a service
Sometimes instead of just getting a reference to the functions from a
dependency service in TK, it is more useful to get a reference to the
service itself (e.g. the object that `satisfies?` the service protocol).
This commit adds a `get-service` function to the main `Service` protocol in TK.