(PE-3657) Added get-shutdown-reason and use by bootstrap test utils
This commit includes a new function on the shutdown service,
get-shutdown-reason, which can be used to get the current shutdown
reason (if one exists) without blocking the current thread until the
reason promise is realized.  The internal get-app-shutdown-reason
function provides a way to get at this new function via a handle to the
TrapperkeeperApp.

This commit also includes changes to the bootstrap testutil functions,
which now make use of the current shutdown reason to translate a failure
to bootstrap without encountering an error into an exception thrown up
the stack.  The bootstrap testutil functions also include overloads with
an optional boolean argument which controls whether or not the failure
to bootstrap should result in the TrapperkeeperApp instance being
returned to the caller (true) vs. an exception being thrown (false, and
the default value).

This commit also factors a few common assertions made across several of
the shutdown tests into a helper function.