Add a test for array-based configuration, add docs
Some of our other supported config formats support arrays for the
values of the data, but it will generally be made available as
arrays of strings.  This commit adds a minor tweak so that
we will call `symbol` on these strings before using them, and
adds a test validating that this type of configuration will
work properly.

The commit also adds documentation for the new setting.