(TK-60) Formalize public function for loading config
If users want to write a command-line tool that doesn't make sense
to boot the full trapperkeeper framework for, but they still want
to leverage trapperkeeper's configuration data, they should be able
to.

Prior to this commit, it was technically possible, but the function
you'd need to have called were marked as private and named poorly
for a public API.  This commit simply formalizes a public `load-config`
function, that users can call w/o needing to boot the full TK
framework.