(maint) add schema validation fixture to all tests
This commit turns on schema validation in all test namespaces via the
standard fixture.

As a result of this, it was also necessary to add additional error
handling in 'hanlde-prismatic-exception!' for the case when schema eagerly
validates the service graph.  In this case, a different exception is
thrown from schema, and it requires a bit more tomfoolery to extract the
list of missing services from it.

An alternative solution to this problem would be to do a manual
validation before instantiating the graph, instead of relying on schema
to throw this exception ... although I haven't yet been able to figure
out how to do that.