(maint) combine 'services' and 'services-internal'
Combine the 'services' and 'services-internal' namespaces.  If memory
serves, these namespaces were initially one, and were separated with the
intention that 'services' would be public API ... I don't think that's
the case anymore, as the only things that would likely be considered
public API at this point are the 'Lifecycle' and 'Service' protocols.
In fact, it might make sense to go a step further and pull those up into
into their own namespace, since they are the only public API outside of
'core'.

Anyway, the combined namespace is by no means large
enough to be unwieldy, and the combination will allow for subsequent
changes to be made that will clean up the code even further.

This entire change was simply:
* copy the contents of 'services-internal', paste into 'services'
* fix up the 'require's/'import's
* ditto for the tests
* delete 'services-internal' and 'services-internal-test'
* move one funky test fixture into its own namespace -
  'testutils.services'