(TK-142) Add support for logback evaluator filters
This commit adds support for logback's EvaluatorFilters.  These
require a dependency that we didn't have before (janino), and
allow really powerful scripting constructs to be specified
in the logback config file.  This will be extremely useful for
support escalations, e.g. in cases where we need to prevent
a problematic log message from bloating the log files on disk.