(TK-202) Use an atom to track all running tk apps
Prior to this commit, we had a single global var that we'd
populate with a reference to the last TK app that was started.
This was intended to make it possible to connect to a running
process via nrepl, and get a reference to the app.

This would cause issues if there were more than one TK app
running in the same process; this commit changes the variable
to be an atom that maintains a vector of all running TK apps.