(MAINT) increase timeout in service test
The recent changes related to HUP support have a test that
sends a HUP signal and then waits for some period of time
to validate that the restart behavior took effect.  The initial
commit of this code appears to be using too low of a timeout
for some of our CI hardware.

This commit bumps the timeout up from 1s to 5s.