(TK-202) Refactor HUP test to call restart-tk-apps
Unfortunately, jenkins clients swallow up HUP signals, so the previous
HUP test consistently fails in jenkins, as the signal never gets to
clojure. This commit addresses this by calling directly into the
function that is invoked by the HUP signal, which is as close as we can
get for the moment.