(TK-334) Improve debug logging around lifecycle / shutdown
This commit adds/improves some debugging statements to clarify what
is happening with lifecycle events inside of TK, and when we shut
down the Clojure agent pool.

To achieve this we needed to provide a bit of schema around the
items that are put onto the core.async channel for lifecycle management.