(TK-330) move shutdown/stop logic to core async worker
This commit moves the "main" shutdown/stop logic off of the main
thread and onto the core.async worker thread.  This ensures that
we don't end up in a situation where we have restarts queued up
and then an error occurs and causes the situation where we have
lifecycle activity going on on the async worker thread while
the main thread is also trying to perform the shutdown tasks.