(TK-331) Tighten lifecycle task schema
Previously, not all lifecycle task messages included a function
to execute, so the schema tolerated a nil for that field.  In
a recent commit, we added a function to the message for 'shutdown',
and thus there are no longer any messages that don't contain a
function.  This allows us to make the schema more restrictive, so
this commit removes the `maybe`.