(TK-331) Move channel cleanup above task-function
This commit moves the core.async channel cleanup in the
shutdown logic up above the call to the task function.  This
way, if an exception were to occur during the task function,
we'd still be assured that the channels had been cleaned up.

I don't think the production code paths should allow for the
possibility of an exception occurring in the shutdown task-function,
but this is safer just in case.