(TK-347) Add support for multiple bootstrap files
This commit adds support for bootstrap configurations to be split up between
multiple .cfg files.

All the .cfg files are essentially concatenated together before tk loads any of
the services in them, so that it appears as one combined bootstrap file.

Multiple bootstrap files are specified on the command line in the same way as
conf are: a comma separated list of files and directories. Because of this,
these changes are backward compatible.

For example, the --bootstrap-config command line option might look like so:
--bootstrap-config /path/to/bootstrap.cfg,/folder/of/bootstrap/files,./another.cfg