(TK-349) Log warning when service can't be found
This commit changes the behavior of TK when a bootstrap file contains a service
entry that can't be resolved. Previous TK would throw an error and fail to
start. Now TK will log a warning and attempt to continue loading the app

This is being added to support workflows where a user might want to add a
service entry to their bootstrap before a TK app actually requires it.

TK continues to throw errors and abort if the service entry is malformed or if
the service is found, but doesn't have a valid service graph