(TK-382) Fix defservice + optional deps
This commit fixes defservice's sadness when specifying optional
dependencies without a protocol. The bug was from use of
clojure.tools.macro/name-with-attributes, which was mistaking the
dependency map for metadata and removing it from the forms passed on to
the service macro itself.

The remediation was to implement our own name-with-attributes that is
smart enough to ignore dependency maps and not blindly accept them as
metadata maps.

This commit also adds tests for the various permutations of input to
defservice.