(TK-405) Support specifying restart-file via a CLI arg
This commit adds the ability for the restart-file setting to be
specified via a CLI arg.  The commit preserves the ability to source the
restart-file setting from the global section in Trapperkeeper
configuration.  In the event that the setting were specified in both
places, the setting used for the CLI arg would be the one used.