(TK-439) Handle exceptions having no type in main
Previously, if an unhandled exception that can be coerced into a map but
that does not have a `:type` key in it were propagated up to
Trapperkeeper's main function, the exception output would be mishandled.
This commit rethrows the original exception if a nil value is found for the
`:type` key in the map, allowing the original exception output to be
displayed properly.