(maint) Add log message for finishing lifecycle function
Previously trapperkeeper would log a message when beginning to run a
lifecycle function, but not when the lifecycle function had returned.
This means that when debugging failures it can be challenging to tell
if a service has died mid-lifecycle function (depending on how much
logging that service has implemented). To alleviate that somewhat, this
commit adds a log message at debug level for when the lifecycle function
has finished.