(maint) Add plugin jars to clojure's base loader
In Java 9 to add a jar to the classpath, we have to add a new
classloader in the hierarchy on the Thread (this is where io/resource
will look) and also to the clojure baseLoader (this is where clojure
will look for actual code). In Java 8 since the base hierarchy was
modifiable we could just change one place and it would be seen
everywhere.