(maint) Remove unnecessary refer-clojure
We previously did not bring in clojure.core's `boolean?` function so that
we could refer kitchensink's. We now bring in all of kitchensink using
the `ks` namespace. This moves us away from partially refering
clojure.core and using the namespace ks/boolean? where appropriate.

A stack trace while debugging dependency upgrades brought me to here,
though on further inspection this was not the cause of the issue.
However, this still seems like a more idiomatic usage, so I thought I
would offer it as a maintenance PR.