(orch-2282) Updates trapperkeeper to use the latest nrepl/nrepl
This is part of a series of changes to support nrepl/nrepl in trapperkeeper. It
needs to be merged after the corresponding changes in clj-parent.

This change requires lein 2.9.0 or later to work; this is when the lein project
changed from the legacy org.clojure/tools.nrepl to the supported nrepl/nrepl
line of development.

This release drops support for JDK7, as nrepl/nrepl has a minimum requirement
of JDK8.