(maint) Update tests for Clojure 1.10
In clojure 1.10 compilation exceptions now seem to be returned as
ex-info rather than the bare exceptions that are required for
`thrown-with-message?`. Consequently this catches the exception (whether
in ex-info form or not) and passes it to `Throwable->map` which will
take either and produce a single representation that we can use.

This update works both on Clojure 1.9 and 1.10.