(PDB-4636) Add top-level-cli ext test to check behavior
Add ext/test/top-level-cli to make sure that running tk with --help,
an invalid otion, and a missing bootstrap file produces the expected
behavior (exit status, etc.):

  lein uberjar
  ext/test/top-level-cli

Add ./tk as a helper that runs trapperkeeper using the standalone
uberjar in target/.

Add ext/test/run-all to run all the external tests, and use it to run
those tests in a new travis job with just jdk11 for now.  Explicitly
list the travis jobs to make them obvious, rather than having to begin
relying on a mixture of implicit and explicit jobs and knowledge of
travis' defaulting rules for the explicit jobs.