(PDB-4636) request-shutdown: support custom exit status/messages
Accept an options map which may request a custom exit status and final
messages like this:

  (request-shutdown
    {:puppetlabs.trapperkepper.core/exit
     {:status 3
      :messages [["Unexpected filesystem error ..." *err*]]}})

Add an initial ext/test/custom-exit-status which can be run directly,
to verify the behavior, and add it to ext/test/run-all.  Use `lein
run` rather than testing the more "final" uberjar to make it easier to
provide a service for the test that tk can find (i.e. in
custom-exit-behavior-test).